requirements.txt
# 指定兼容版本
protobuf==3.20.3
typing-extensions==4.9.0
requests==2.31.0
packaging==23.2
grpcio==1.60.0

# 主要依赖
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-chroma>=0.1.0
dashscope>=1.17.0
chromadb>=0.4.0
pypdf>=3.17.0
pandas>=2.0.0
openpyxl>=3.1.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
streamlit>=1.28.0